代码优化

+++ internal/conf/conf.go
@@ -66,7 +66,7 @@ type Ranking struct {
 	PostTimeKey       string  `mapstructure:"post_time_key"`
 	PostScoreKey      string  `mapstructure:"post_score_key"`
 	PostVotedPrefix   string  `mapstructure:"post_voted_prefix"`
-	PostVotingPeriod  int64   `mapstructure:"post_voting_period"`
+	PostVotingPeriod  float64 `mapstructure:"post_voting_period"`
 	PostVoteUnitScore float64 `mapstructure:"post_vote_unit_score"`
 }
